requirements.txt
streamlit
mibian
pyyaml
pandas
numpy
pyzmq
textual
rich
requests
